Fixed errors in document
This is a botched job of implementing our translations. Either you did not give us good instruction how to provide you with the translation or you did not have a look at the files you received.

Take a look at the edits and see if you can see recurring mistakes that are also in many other translations.